added post request to register a new user and to login a user. Updated BaseEntity and UserEntity, as well as response and request models. Update id to Long

+++ src/main/java/com/ag/trigger/payload/response/JwtResponse.java
@@ -9,12 +9,12 @@ public class JwtResponse {
 
     private String token;
     private String type = "Bearer";
-    private Integer id;
+    private Long id;
     private String username;
     private String email;
     private List<String> roles;
 
-    public JwtResponse(String token, Integer id, String username, String email, List<String> roles) {
+    public JwtResponse(String token, Long id, String username, String email, List<String> roles) {
         this.token = token;
         this.id = id;
         this.username = username;
